// cgoml-ui/src/app/features/models/messaging/views/full/InboundMessageFullView.ts
import { InboundMessageForm } from 'src/app/features/models/messaging/forms/InboundMessageForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class InboundMessageFullView {

id: number;
form: InboundMessageForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}